+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 ========================
 
 New functionality added in a backwards-compatible manner:
-* [OLMIS-3222](https://openlmis.atlassian.net/browse/OLMIS-3222): Added start and end dates parameters to the order view screen
+* [OLMIS-3222](https://openlmis.atlassian.net/browse/OLMIS-3222): Added period start and end dates parameters to the order view screen
 
 Bug fixes: